+++ savedsearch-comparator/savedsearchcomparator-web/src/main/java/oracle/sysman/emaas/platform/savedsearch/comparator/ws/rest/comparator/rows/entities/SavedSearchCategoryRowEntity.java
@@ -10,6 +10,8 @@
 
 package oracle.sysman.emaas.platform.savedsearch.comparator.ws.rest.comparator.rows.entities;
 
+import java.math.BigInteger;
+
 import org.codehaus.jackson.annotate.JsonProperty;
 
 /**
@@ -18,7 +20,7 @@ import org.codehaus.jackson.annotate.JsonProperty;
 public class SavedSearchCategoryRowEntity implements RowEntity
 {
 	@JsonProperty("CATEGORY_ID")
-	private Long categoryId;
+	private BigInteger categoryId;
 
 	@JsonProperty("DESCRIPTION")
 	private String description;
@@ -45,7 +47,7 @@ public class SavedSearchCategoryRowEntity implements RowEntity
 	private String owner;
 
 	@JsonProperty("DELETED")
-	private Long deleted;
+	private BigInteger deleted;
 
 	@JsonProperty("PROVIDER_NAME")
 	private String providerName;
@@ -215,7 +217,7 @@ public class SavedSearchCategoryRowEntity implements RowEntity
 	/**
 	 * @return the categoryId
 	 */
-	public Long getCategoryId()
+	public BigInteger getCategoryId()
 	{
 		return categoryId;
 	}
@@ -223,7 +225,7 @@ public class SavedSearchCategoryRowEntity implements RowEntity
 	/**
 	 * @return the deleted
 	 */
-	public Long getDeleted()
+	public BigInteger getDeleted()
 	{
 		return deleted;
 	}
@@ -371,7 +373,7 @@ public class SavedSearchCategoryRowEntity implements RowEntity
 	 * @param categoryId
 	 *            the categoryId to set
 	 */
-	public void setCategoryId(Long categoryId)
+	public void setCategoryId(BigInteger categoryId)
 	{
 		this.categoryId = categoryId;
 	}
@@ -380,7 +382,7 @@ public class SavedSearchCategoryRowEntity implements RowEntity
 	 * @param deleted
 	 *            the deleted to set
 	 */
-	public void setDeleted(Long deleted)
+	public void setDeleted(BigInteger deleted)
 	{
 		this.deleted = deleted;
 	}
